requirements.txt
requests
mistralai
openai
speechrecognition
edge-tts
vosk
playgrounds
python-dotenv
yt_dlp
